Moves tdyrichards.h, tdyrockproperties.h, and tdyth.h to include/private; and updates TDyIO
